Updates source code for Win32

+++ examples/beauty_client_async_ssl.cpp
@@ -37,7 +37,7 @@ int main(int argc, char* argv[])
     int step        = count / 10;
     if (!step) step = 1;
 
-    int message_size = 0;
+    size_t message_size = 0;
     std::atomic_int success = 0;
     std::atomic_int failure  = 0;
 
@@ -99,6 +99,4 @@ int main(int argc, char* argv[])
             std::chrono::duration<double>>(stop - start).count();
     std::cout << "Total duration:      " << delay << " seconds" << std::endl;
     std::cout  << std::fixed << "Requests per seconds: " << (count / delay) << " [#/sec]" << std::endl;
-
-
 }
